fastapi-backend/requirements.txt
anyio==3.7.0
cffi==1.15.1
click==8.1.3
cryptography==41.0.2
ecdsa==0.18.0
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
mysqlclient==2.2.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
